database in eigene Datei ausgelagert
