+++ Client_Sub/Transmit.cpp
@@ -1,8 +1,6 @@
 #include <RCSwitch.h>
-#include <Effortless_SPIFFS.h>
-#include <ArduinoJson.h>
 #include "Transmit.h"
-
+#include "database.h"
 
 RCSwitch mySwitch2 = RCSwitch();
 
@@ -21,23 +19,8 @@ RCSwitch mySwitch2 = RCSwitch();
 #define repeat 3
 #define transmitterPin D3
 
-#define DB_PATH "/database.json"
 
-DynamicJsonDocument database(8192);
-eSPIFFS fileSystem;
 bool isRecordingCodes = false;
-
-void read_database() {
-  String db_string;
-  if (fileSystem.openFromFile(DB_PATH, db_string)) {
-    Serial.print("Successfully read file and parsed data: ");
-    deserializeJson(database, db_string);
-    Serial.println(db_string);
-  } else {
-    // DB is new
-    Serial.println("Created a new database.");
-  }
-}
 
 void transmit_setup() {
 
@@ -70,33 +53,3 @@ void transmit(String name, String state){
    mySwitch2.sendTriState(database[name][state][1]);
 }
 
-
-void saveDatabase() {
-  String db_string;
-  serializeJson(database, db_string);
-  Serial.print("Saving database: ");
-  Serial.println(db_string);
-  if (fileSystem.saveToFile(DB_PATH, db_string)) {
-      Serial.println("Saving complete.");
-  } else {
-    Serial.println("Error saving database.");
-  }
-}
-
-void addToDatabase(String name, String state, String code) {
-  if (state != "off" && state != "on") return;
-
-  if (!database.containsKey(name)) {
-    // Name is new
-    database[name]["on"][0] = 1; // next code will be placed at this index
-    database[name]["off"][0] = 1;
-    database[name][state][1] = code;
-    database[name][state][0] = 2;
-    serializeJson(database, Serial); // TODO: remove in production
-    return;
-  }
-
-  database[name][state][database[name][state][0]] = code;
-  database[name][state][0] = database[name][state][0].as<int32_t>() + 1;
-  serializeJson(database, Serial);
-}
