+++ Client_Sub/Transmit.h
@@ -4,8 +4,6 @@
 
 void transmit_setup();
 void transmit(String name, String state);
-void saveDatabase();
-void addToDatabase(String name, String state, String code);
 
 
 #endif
